pyproject.toml
[project]
name = "docsumo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datauri>=1.0.0",
    "easyocr>=1.7.2",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "johnsnowlabs>=6.0.0",
    "llmwhisperer-client>=2.3.0",
    "matplotlib>=3.10.1",
    "mistralai>=1.7.0",
    "paddleocr>=2.10.0",
    "paddlepaddle>=3.0.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "tesseract>=0.1.3",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
]
